JHipster/src/test/javascript/spec/app/entities/test-category/test-category-update.component.spec.ts
/* tslint:disable max-line-length */
import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { Observable, of } from 'rxjs';

import { JtTestingTestModule } from '../../../test.module';
import { TestCategoryUpdateComponent } from 'app/entities/test-category/test-category-update.component';
import { TestCategoryService } from 'app/entities/test-category/test-category.service';
import { TestCategory } from 'app/shared/model/test-category.model';

describe('Component Tests', () => {
    describe('TestCategory Management Update Component', () => {
        let comp: TestCategoryUpdateComponent;
        let fixture: ComponentFixture<TestCategoryUpdateComponent>;
        let service: TestCategoryService;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [JtTestingTestModule],
                declarations: [TestCategoryUpdateComponent]
            })
                .overrideTemplate(TestCategoryUpdateComponent, '')
                .compileComponents();

            fixture = TestBed.createComponent(TestCategoryUpdateComponent);
            comp = fixture.componentInstance;
            service = fixture.debugElement.injector.get(TestCategoryService);
        });

        describe('save', () => {
            it(
                'Should call update service on save for existing entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new TestCategory(123);
                    spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.testCategory = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.update).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );

            it(
                'Should call create service on save for new entity',
                fakeAsync(() => {
                    // GIVEN
                    const entity = new TestCategory();
                    spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
                    comp.testCategory = entity;
                    // WHEN
                    comp.save();
                    tick(); // simulate async

                    // THEN
                    expect(service.create).toHaveBeenCalledWith(entity);
                    expect(comp.isSaving).toEqual(false);
                })
            );
        });
    });
});
